+++ app/Http/Controllers/RolesController.php
@@ -8,15 +8,19 @@ use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Models\Permission;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class RolesController extends Controller
 {
     public function index() {
+        abort_if(Gate::denies('roles_management'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $roles = Role::all();
         return view('roles.index', compact('roles'));
     }
 
     public function create() {
+        abort_if(Gate::denies('roles_management'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permissions = Permission::pluck('title', 'id');
         return view('roles.create', compact('permissions'));
     }
@@ -28,10 +32,12 @@ class RolesController extends Controller
     }
 
     public function show(Role $role) {
+        abort_if(Gate::denies('roles_management'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('roles.show', compact('role'));
     }
 
     public function edit(Role $role) {
+        abort_if(Gate::denies('roles_management'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permissions = Permission::pluck('title', 'id');
         $role->load('permissions');
 
@@ -45,6 +51,7 @@ class RolesController extends Controller
     }
 
     public function destroy(Role $role) {
+        abort_if(Gate::denies('roles_management'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $role->delete();
         return redirect()->route('roles.index');
     }
